Keep reference to metadata before regenerating images.
Fixes #136

Starting from WP 5.3 "sizes" attribute on metadata is reset to empty
when passed to `intermediate_image_sizes_advanced` filter.
The fix saves temporarily the metadata for the original file before
calling the `intermediate_image_sizes_advanced` filter, so we can keep
the same logic that was being applied before.